CR: Allow leave Profile 0.5 adjustment

+++ app/models/LeaveProfileModel.scala
@@ -29,7 +29,7 @@ case class LeaveProfile (
 case class LeaveProfileCalculation (
     ent: Int,
     ear: Double,
-    adj: Int,
+    adj: Double,
     uti: Double,
     cf: Double,
     cfuti: Double,
@@ -115,7 +115,7 @@ object LeaveProfileModel {
       LeaveProfileCalculation(
           doc.getAs[Int]("ent").get,
           doc.getAs[Double]("ear").get,
-          doc.getAs[Int]("adj").get,
+          doc.getAs[Double]("adj").get,
           doc.getAs[Double]("uti").get,
           doc.getAs[Double]("cf").get,
           doc.getAs[Double]("cfuti").get,
@@ -239,7 +239,7 @@ object LeaveProfileModel {
       pid = "",
       pn = "",
       lt = "",
-      cal = LeaveProfileCalculation (ent = 0, ear = 0.0, adj = 0, uti = 0.0, cf = 0.0, cfuti = 0.0, cfexp = 0.0, bal = 0.0),
+      cal = LeaveProfileCalculation (ent = 0, ear = 0.0, adj = 0.0, uti = 0.0, cf = 0.0, cfuti = 0.0, cfexp = 0.0, bal = 0.0),
       me = LeaveProfileMonthEarn(jan=0.0, feb=0.0, mar=0.0, apr=0.0, may=0.0, jun=0.0, jul=0.0, aug=0.0, sep = 0.0, oct=0.0, nov=0.0, dec=0.0),
       set_ent = Entitlement(e1=0, e1_s=0, e1_cf=0, e2=0, e2_s=0, e2_cf=0, e3=0, e3_s=0, e3_cf=0, e4=0, e4_s=0, e4_cf=0, e5=0, e5_s=0, e5_cf=0),
       sys=None
